CC-16029: Quality assurance

+++ src/SprykerEco/Zed/Unzer/Business/ApiAdapter/Mapper/UnzerGetPaymentMapper.php
@@ -65,7 +65,7 @@ class UnzerGetPaymentMapper implements UnzerGetPaymentMapperInterface
         foreach ($unzerApiGetPaymentResponseTransfer->getTransactions() as $unzerApiTransactionTransfer) {
             $unzerTransactionTransfer = (new UnzerTransactionTransfer())
                 ->fromArray($unzerApiTransactionTransfer->toArray(), true)
-                ->setAmount((int)$unzerApiTransactionTransfer->getAmount() * UnzerConstants::INT_TO_FLOAT_DIVIDER);
+                ->setAmount((float)$unzerApiTransactionTransfer->getAmount() * UnzerConstants::INT_TO_FLOAT_DIVIDER);
 
             $unzerTransactionTransfers->append($unzerTransactionTransfer);
         }
